requirements.txt
python-dotenv==1.0.0
asyncpg==0.27.0
alembic==1.11.0
SQLAlchemy==2.0.13
fastapi==0.95.1
uvicorn==0.22.0
aiohttp==3.8.4
httpx==0.24.1
pytest==7.3.1
pytest-asyncio==0.21.0